eureka_visual/requirements.txt
gymnasium[mujoco]
stable-baselines3
transformers
torch
pillow
numpy
requests
psutil
